Move cluster details worker nodes widget to worker nodes folder
